requirements.txt
Flask==3.0.3
dash==2.18.2
dash-bootstrap-components==1.6.0
flask-socketio==5.3.6
eventlet==0.36.1
paho-mqtt==2.1.0
influxdb-client==1.48.0
python-dotenv==1.0.1
gunicorn==22.0.0

